owlDataHandler refactoring
