+++ onto-viewer-core/src/main/java/org/edmcouncil/spec/ontoviewer/core/ontology/DetailsManager.java
@@ -16,7 +16,15 @@ import org.edmcouncil.spec.ontoviewer.core.model.details.OwlGroupedDetails;
 import org.edmcouncil.spec.ontoviewer.core.model.details.OwlListDetails;
 import org.edmcouncil.spec.ontoviewer.core.model.module.OntologyModule;
 import org.edmcouncil.spec.ontoviewer.core.model.property.OwlAnnotationPropertyValue;
-import org.edmcouncil.spec.ontoviewer.core.ontology.data.OwlDataHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.data.AnnotationsDataHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.classes.ClassHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.module.ModuleHelper;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.metadata.MetadataHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.module.ModuleHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.data.DataPropertyHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.data.DataTypeHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.individual.IndividualHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.data.ObjectDataHandler;
 import org.edmcouncil.spec.ontoviewer.core.ontology.generator.DescriptionGenerator;
 import org.edmcouncil.spec.ontoviewer.core.service.ChangerIriToLabelService;
 import org.semanticweb.owlapi.model.EntityType;
@@ -40,20 +48,42 @@ public class DetailsManager {
 
   private final ApplicationConfigurationService applicationConfigurationService;
   private final OntologyManager ontologyManager;
-  private final OwlDataHandler dataHandler;
+  private final ClassHandler particularClassHandler;
   private final ChangerIriToLabelService changerIriToLabelService;
   private final DescriptionGenerator descriptionGenerator;
+  private final ModuleHandler moduleHandler;
+  private final ModuleHelper moduleHelper;
+  private final AnnotationsDataHandler annotationsDataHandler;
+  private final DataTypeHandler particularDataTypeHandler;
+  private final ObjectDataHandler particularObjectPropertyHandler;
+  private final DataPropertyHandler particularDataPropertyHandler;
+  private final IndividualHandler particularIndividualHandler;
+  private final MetadataHandler metadataHandler;
 
   public DetailsManager(ApplicationConfigurationService applicationConfigurationService,
       OntologyManager ontologyManager,
-      OwlDataHandler dataHandler,
+      ClassHandler particularClassHandler,
       ChangerIriToLabelService changerIriToLabelService,
-      DescriptionGenerator descriptionGenerator) {
+      DescriptionGenerator descriptionGenerator, ModuleHandler moduleHandler,
+      ModuleHelper moduleHelper,
+      AnnotationsDataHandler annotationsDataHandler,
+      DataTypeHandler particularDataTypeHandler,
+      ObjectDataHandler particularObjectPropertyHandler,
+      DataPropertyHandler particularDataPropertyHandler,
+      IndividualHandler particularIndividualHandler, MetadataHandler metadataHandler) {
     this.applicationConfigurationService = applicationConfigurationService;
     this.ontologyManager = ontologyManager;
-    this.dataHandler = dataHandler;
+    this.particularClassHandler = particularClassHandler;
     this.changerIriToLabelService = changerIriToLabelService;
     this.descriptionGenerator = descriptionGenerator;
+    this.moduleHandler = moduleHandler;
+    this.moduleHelper = moduleHelper;
+    this.annotationsDataHandler = annotationsDataHandler;
+    this.particularDataTypeHandler = particularDataTypeHandler;
+    this.particularObjectPropertyHandler = particularObjectPropertyHandler;
+    this.particularDataPropertyHandler = particularDataPropertyHandler;
+    this.particularIndividualHandler = particularIndividualHandler;
+    this.metadataHandler = metadataHandler;
   }
 
   public OWLOntology getOntology() {
@@ -67,19 +97,23 @@ public class DetailsManager {
     OwlListDetails result = null;
 
     if (entityType == EntityType.CLASS) {
-      result = dataHandler.handleParticularClass(owlEntity.asOWLClass());
+      result = particularClassHandler.handleParticularClass(owlEntity.asOWLClass());
     } else if (entityType == EntityType.NAMED_INDIVIDUAL) {
-      result = dataHandler.handleParticularIndividual(owlEntity.asOWLNamedIndividual());
+      result = particularIndividualHandler.handleParticularIndividual(
+          owlEntity.asOWLNamedIndividual());
     } else if (entityType == EntityType.OBJECT_PROPERTY) {
-      result = dataHandler.handleParticularObjectProperty(owlEntity.asOWLObjectProperty());
+      result = particularObjectPropertyHandler.handleParticularObjectProperty(
+          owlEntity.asOWLObjectProperty());
     } else if (entityType == EntityType.DATA_PROPERTY) {
-      result = dataHandler.handleParticularDataProperty(owlEntity.asOWLDataProperty());
+      result = particularDataPropertyHandler.handleParticularDataProperty(
+          owlEntity.asOWLDataProperty());
     } else if (entityType == EntityType.DATATYPE) {
-      result = dataHandler.handleParticularDatatype(owlEntity.asOWLDatatype());
+      result = particularDataTypeHandler.handleParticularDatatype(owlEntity.asOWLDatatype());
     }
 
     if (result == null) {
-      throw new NotFoundElementInOntologyException(String.format(NOT_FOUND_ENTITY_MESSAGE, entityIri.toString()));
+      throw new NotFoundElementInOntologyException(
+          String.format(NOT_FOUND_ENTITY_MESSAGE, entityIri.toString()));
     }
 
     result.setIri(entityIri.toString());
@@ -90,11 +124,11 @@ public class DetailsManager {
         .isLocationInModulesEnabled();
     if (locationInModulesEnabled) {
       result.setLocationInModules(
-          dataHandler.getElementLocationInModules(
+          moduleHelper.getElementLocationInModules(
               entityIri.getIRIString()));
     }
 
-    result.setMaturityLevel(dataHandler.getMaturityLevel(entityIri));
+    result.setMaturityLevel(moduleHelper.getMaturityLevel(entityIri));
 
     ConfigurationData coreConfiguration = applicationConfigurationService.getConfigurationData();
     if (applicationConfigurationService.hasConfiguredGroups()) {
@@ -114,25 +148,27 @@ public class DetailsManager {
 
     // If '/' is at the end of the URL, we extract the ontology metadata
     if (iriString.endsWith("/")) {
-      result = dataHandler.handleOntologyMetadata(iri);
+      result = metadataHandler.handleOntologyMetadata(iri);
     } else {
       result = getEntityListDetails(iri);
     }
 
     if (result == null) {
-      throw new NotFoundElementInOntologyException(String.format(NOT_FOUND_ENTITY_MESSAGE, iriString));
+      throw new NotFoundElementInOntologyException(
+          String.format(NOT_FOUND_ENTITY_MESSAGE, iriString));
     }
 
     return setGroupedDetailsIfEnabled(iriString, result);
   }
 
   public List<OntologyModule> getAllModulesData() {
-    return dataHandler.getAllModules();
+    return moduleHelper.getAllModules();
   }
 
-  public OwlDetails getEntityDetailsByIri(String iriString) throws NotFoundElementInOntologyException {
+  public OwlDetails getEntityDetailsByIri(String iriString)
+      throws NotFoundElementInOntologyException {
     IRI iri = IRI.create(iriString);
-    OwlListDetails resultDetails = dataHandler.handleOntologyMetadata(iri);
+    OwlListDetails resultDetails = metadataHandler.handleOntologyMetadata(iri);
     if (resultDetails == null) {
       resultDetails = getEntityListDetails(iri);
 
@@ -140,7 +176,6 @@ public class DetailsManager {
         throw new NotFoundElementInOntologyException(String.format(NOT_FOUND_ENTITY_MESSAGE, iri));
       }
     }
-
     return setGroupedDetailsIfEnabled(iriString, resultDetails);
   }
 
@@ -149,7 +184,7 @@ public class DetailsManager {
         .getToolkitConfig()
         .isLocationInModulesEnabled();
     if (locationInModulesEnabled) {
-      result.setLocationInModules(dataHandler.getElementLocationInModules(iriString));
+      result.setLocationInModules(moduleHelper.getElementLocationInModules(iriString));
     }
 
     ConfigurationData configurationData = applicationConfigurationService.getConfigurationData();
@@ -168,28 +203,29 @@ public class DetailsManager {
     OwlListDetails result = null;
 
     if (ontologyManager.getOntology().containsClassInSignature(iri, INCLUDED)) {
-      result = dataHandler.handleParticularClass(iri);
+      result = particularClassHandler.handleParticularClass(iri);
     } else if (ontologyManager.getOntology().containsDataPropertyInSignature(iri, INCLUDED)) {
-      result = dataHandler.handleParticularDataProperty(iri);
+      result = particularDataPropertyHandler.handleParticularDataProperty(iri);
     } else if (ontologyManager.getOntology().containsObjectPropertyInSignature(iri, INCLUDED)) {
-      result = dataHandler.handleParticularObjectProperty(iri);
+      result = particularObjectPropertyHandler.handleParticularObjectProperty(iri);
     } else if (ontologyManager.getOntology().containsIndividualInSignature(iri, INCLUDED)) {
-      result = dataHandler.handleParticularIndividual(iri);
+      result = particularIndividualHandler.handleParticularIndividual(iri);
     } else if (ontologyManager.getOntology().containsDatatypeInSignature(iri, INCLUDED)) {
-      result = dataHandler.handleParticularDatatype(iri);
+      result = particularDataTypeHandler.handleParticularDatatype(iri);
     } else if (ontologyManager.getOntology().containsAnnotationPropertyInSignature(iri, INCLUDED)) {
-      result = dataHandler.handleParticularAnnotationProperty(iri, getOntology());
+      result = annotationsDataHandler.handleParticularAnnotationProperty(iri, getOntology());
     }
 
     if (result != null) {
       result.setIri(iri.toString());
-      result.setMaturityLevel(dataHandler.getMaturityLevel(iri));
+      result.setMaturityLevel(moduleHelper.getMaturityLevel(iri));
     }
 
     return result;
   }
 
-  private OwlGroupedDetails groupDetails(OwlListDetails owlDetails, ConfigurationData configurationData) {
+  private OwlGroupedDetails groupDetails(OwlListDetails owlDetails,
+      ConfigurationData configurationData) {
     var groupedDetails = new OwlGroupedDetails();
     var groups = configurationData.getGroupsConfig().getGroups();
 
@@ -228,7 +264,6 @@ public class DetailsManager {
         return groupEntry.getKey();
       }
     }
-
     return null;
   }
