+++ onto-viewer-core/src/main/java/org/edmcouncil/spec/ontoviewer/core/ontology/data/handler/LicenseHandler.java
@@ -4,7 +4,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.stream.Collector;
 import java.util.stream.Collectors;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.model.ConfigurationData.ApplicationConfig;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.service.ApplicationConfigurationService;
@@ -13,6 +12,7 @@ import org.edmcouncil.spec.ontoviewer.core.model.PropertyValue;
 import org.edmcouncil.spec.ontoviewer.core.model.property.OwlAnnotationPropertyValue;
 import org.edmcouncil.spec.ontoviewer.core.model.property.OwlDetailsProperties;
 import org.edmcouncil.spec.ontoviewer.core.ontology.OntologyManager;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.module.ModuleHandler;
 import org.semanticweb.owlapi.model.IRI;
 import org.semanticweb.owlapi.model.OWLAnnotation;
 import org.semanticweb.owlapi.model.OWLOntology;
@@ -20,7 +20,6 @@ import org.semanticweb.owlapi.model.OWLOntologyManager;
 import org.springframework.stereotype.Service;
 
 /**
- *
  * @author patrycja.miazek (patrycja.miazek@makolab.com)
  */
 @Service
@@ -30,9 +29,12 @@ public class LicenseHandler {
   private final ApplicationConfig applicationConfig;
   private final ModuleHandler moduleHandler;
 
-  public LicenseHandler(OntologyManager ontologyManager, ApplicationConfigurationService applicationConfigurationService, ModuleHandler moduleHandler) {
+  public LicenseHandler(OntologyManager ontologyManager,
+      ApplicationConfigurationService applicationConfigurationService,
+      ModuleHandler moduleHandler) {
     this.ontologyManager = ontologyManager;
-    this.applicationConfig = applicationConfigurationService.getConfigurationData().getApplicationConfig();
+    this.applicationConfig = applicationConfigurationService.getConfigurationData()
+        .getApplicationConfig();
     this.moduleHandler = moduleHandler;
   }
 
@@ -47,14 +49,16 @@ public class LicenseHandler {
 
       if (ontologyIriOptional.isPresent()) {
         var currentOntologyIri = ontologyIriOptional.get();
-        var createIri = IRI.create(ontologyIri.getIRIString().substring(0, ontologyIri.getIRIString().length() - 1));
+        var createIri = IRI.create(
+            ontologyIri.getIRIString().substring(0, ontologyIri.getIRIString().length() - 1));
 
         if ((currentOntologyIri.equals(ontologyIri)
             || currentOntologyIri.equals(createIri))
             && !visitedOntologies.contains(currentOntologyIri.toString())) {
           visitedOntologies.add(currentOntologyIri.toString());
 
-          for (OWLAnnotation annotation : currentOntology.annotations().collect(Collectors.toList())) {
+          for (OWLAnnotation annotation : currentOntology.annotations()
+              .collect(Collectors.toList())) {
 
             for (String licenseIri : applicationConfig.getLicense()) {
               if (licenseIri.equals(annotation.getProperty().getIRI().toString())) {
@@ -82,9 +86,10 @@ public class LicenseHandler {
     }
     return null;
   }
-  
+
   public boolean isLicenseExist(OwlDetailsProperties<PropertyValue> owlDetailsProperties) {
-    for (Map.Entry<String, List<PropertyValue>> entry : owlDetailsProperties.getProperties().entrySet()) {
+    for (Map.Entry<String, List<PropertyValue>> entry : owlDetailsProperties.getProperties()
+        .entrySet()) {
       if (applicationConfig.getLicense().contains(entry.getKey())) {
         return true;
       }
