+++ onto-viewer-core/src/main/java/org/edmcouncil/spec/ontoviewer/core/ontology/data/handler/module/ModuleHandler.java
@@ -1,4 +1,4 @@
-package org.edmcouncil.spec.ontoviewer.core.ontology.data.handler;
+package org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.module;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -22,6 +22,7 @@ import org.edmcouncil.spec.ontoviewer.core.model.module.OntologyModule;
 import org.edmcouncil.spec.ontoviewer.core.model.property.OwlDetailsProperties;
 import org.edmcouncil.spec.ontoviewer.core.model.property.OwlListElementIndividualProperty;
 import org.edmcouncil.spec.ontoviewer.core.ontology.OntologyManager;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.individual.IndividualDataHelper;
 import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.maturity.MaturityLevel;
 import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.maturity.MaturityLevelFactory;
 import org.edmcouncil.spec.ontoviewer.core.ontology.data.label.LabelProvider;
@@ -45,8 +46,16 @@ import org.springframework.stereotype.Service;
 public class ModuleHandler {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ModuleHandler.class);
+
+  private final OntologyManager ontologyManager;
+  private final IndividualDataHelper individualDataHandler;
+  private final LabelProvider labelProvider;
+  private final MaturityLevelFactory maturityLevelFactory;
+
+
   private static final String ONTOLOGY_IRI_GROUP_NAME = "ontologyIri";
-  private static final Pattern ONTOLOGY_IRI_PATTERN = Pattern.compile("(?<ontologyIri>.*\\/)[^/]+$");
+  private static final Pattern ONTOLOGY_IRI_PATTERN = Pattern.compile(
+      "(?<ontologyIri>.*\\/)[^/]+$");
   private static final String HAS_PART_IRI = "http://purl.org/dc/terms/hasPart";
   private static final String INSTANCE_KEY = ViewerIdentifierFactory.createId(
       ViewerIdentifierFactory.Type.function,
@@ -54,28 +63,26 @@ public class ModuleHandler {
   private static final Pattern URL_PATTERN
       = Pattern.compile("^(https?|ftp|file)://[-a-zA-Z0-9+&@#/%?=~_|!:,.;]*[-a-zA-Z0-9+&@#/%=~_|]");
   private final boolean automaticCreationOfModules;
-  private final OntologyManager ontologyManager;
-  private final IndividualDataHandler individualDataHandler;
-  private final LabelProvider labelProvider;
   private final Set<IRI> ontologiesToIgnoreWhenGeneratingModules;
   private final Set<Pattern> ontologyModuleIgnorePatterns;
   private final OWLAnnotationProperty hasPartAnnotation;
   // Cache for ontologies' maturity level
   private final Map<IRI, MaturityLevel> maturityLevelsCache = new ConcurrentHashMap<>();
   private final String moduleClassIri;
-  private final MaturityLevelFactory maturityLevelFactory;
+
 
   private List<OntologyModule> modules;
   private Map<IRI, OntologyModule> modulesMap;
 
-  public ModuleHandler(OntologyManager ontologyManager,
-                       IndividualDataHandler individualDataHandler,
-                       LabelProvider labelProvider,
-                       ApplicationConfigurationService applicationConfigurationService,
-                       MaturityLevelFactory maturityLevelFactory) {
+  public ModuleHandler(
+      OntologyManager ontologyManager, IndividualDataHelper individualDataHandler,
+      LabelProvider labelProvider,
+      MaturityLevelFactory maturityLevelFactory,
+      ApplicationConfigurationService applicationConfigurationService) {
     this.ontologyManager = ontologyManager;
     this.individualDataHandler = individualDataHandler;
     this.labelProvider = labelProvider;
+    this.maturityLevelFactory = maturityLevelFactory;
     this.automaticCreationOfModules = applicationConfigurationService.getConfigurationData()
         .getOntologiesConfig()
         .getAutomaticCreationOfModules();
@@ -94,9 +101,8 @@ public class ModuleHandler {
         .stream()
         .map(Pattern::compile)
         .collect(Collectors.toSet());
-    this.maturityLevelFactory = maturityLevelFactory;
-
-    this.hasPartAnnotation = OWLManager.getOWLDataFactory().getOWLAnnotationProperty(IRI.create(HAS_PART_IRI));
+    this.hasPartAnnotation = OWLManager.getOWLDataFactory()
+        .getOWLAnnotationProperty(IRI.create(HAS_PART_IRI));
     this.moduleClassIri = applicationConfigurationService
         .getConfigurationData()
         .getOntologiesConfig()
@@ -112,20 +118,20 @@ public class ModuleHandler {
       if (moduleClassIri != null) {
         IRI moduleIri = IRI.create(moduleClassIri);
         Optional<OWLClass> moduleClassOptional = ontology
-                .classesInSignature(Imports.INCLUDED)
-                .filter(c -> c.getIRI().equals(moduleIri))
-                .findFirst();
+            .classesInSignature(Imports.INCLUDED)
+            .filter(c -> c.getIRI().equals(moduleIri))
+            .findFirst();
 
         if (moduleClassOptional.isPresent()) {
           OwlDetailsProperties<PropertyValue> moduleClass
-                  = individualDataHandler.handleClassIndividuals(ontology, moduleClassOptional.get());
+              = individualDataHandler.handleClassIndividuals(ontology, moduleClassOptional.get());
 
           if (!moduleClass.getProperties().isEmpty()) {
             Set<String> modulesIris = moduleClass.getProperties().get(INSTANCE_KEY)
-                    .stream()
-                    .map(OwlListElementIndividualProperty.class::cast)
-                    .map(individualProperty -> individualProperty.getValue().getIri())
-                    .collect(Collectors.toSet());
+                .stream()
+                .map(OwlListElementIndividualProperty.class::cast)
+                .map(individualProperty -> individualProperty.getValue().getIri())
+                .collect(Collectors.toSet());
 
             List<String> rootModulesIris = getRootModulesIris(modulesIris);
 
@@ -244,7 +250,7 @@ public class ModuleHandler {
     return maturityLevelFactory.notSet();
   }
 
-  void updateModules() {
+  public void updateModules() {
     this.modules = null;
     // If modules is empty is auto generated again while get
     getModules();
@@ -278,7 +284,6 @@ public class ModuleHandler {
       }
       return null;
     }
-
     return elementIri;
   }
 
@@ -302,7 +307,6 @@ public class ModuleHandler {
         }
       }
     }
-
     return Optional.empty();
   }
 
@@ -316,17 +320,16 @@ public class ModuleHandler {
       checkAndCompleteMaturityLevel(subModule);
       levels.add(subModule.getMaturityLevel());
     }
-
     module.setMaturityLevel(chooseMaturityLevel(levels));
   }
 
-  private boolean isMaturityLevelInSet(MaturityLevel maturityLevel, List<MaturityLevel> maturityLevels) {
+  private boolean isMaturityLevelInSet(MaturityLevel maturityLevel,
+      List<MaturityLevel> maturityLevels) {
     var maturityLevelOptional = maturityLevelFactory.getByIri(maturityLevel.getIri());
     if (maturityLevelOptional.isPresent()) {
       var maturityLevelDefinition = maturityLevelOptional.get();
       return maturityLevels.contains(maturityLevelDefinition);
     }
-
     return false;
   }
 
@@ -357,7 +360,8 @@ public class ModuleHandler {
           if (ontologyIriOptional.isPresent()) {
             var ontologyIri = ontologyIriOptional.get();
 
-            var ontologyLoadedDirectly = ontologyManager.getIriToPathMapping().get(ontologyIri) != null;
+            var ontologyLoadedDirectly =
+                ontologyManager.getIriToPathMapping().get(ontologyIri) != null;
 
             if (ontologyLoadedDirectly) {
               return checkIfShouldAddModuleForOntology(modulesIris, ontologyIri);
@@ -403,7 +407,8 @@ public class ModuleHandler {
           }
         }
       } else {
-        var ontologyPathWithoutFilePrefix = PathUtils.getPathWithoutFilePrefix(ontologyPath.toString());
+        var ontologyPathWithoutFilePrefix = PathUtils.getPathWithoutFilePrefix(
+            ontologyPath.toString());
         var ontologyFileName = ontologyPathWithoutFilePrefix.getFileName();
         for (Pattern pattern : ontologyModuleIgnorePatterns) {
           var match = pattern.matcher(ontologyFileName.toString());
@@ -413,7 +418,6 @@ public class ModuleHandler {
         }
       }
     }
-
     return !modulesIris.contains(ontologyIri.toString());
   }
 
@@ -437,7 +441,8 @@ public class ModuleHandler {
     return ontologyModules;
   }
 
-  private void addSubModules(List<OntologyModule> subModule, Map<IRI, OntologyModule> ontologyModules) {
+  private void addSubModules(List<OntologyModule> subModule,
+      Map<IRI, OntologyModule> ontologyModules) {
     for (OntologyModule ontologyModule : subModule) {
       ontologyModules.put(IRI.create(ontologyModule.getIri()), ontologyModule);
       addSubModules(ontologyModule.getSubModule(), ontologyModules);
@@ -466,7 +471,6 @@ public class ModuleHandler {
       String s = annotation.annotationValue().toString();
       result.add(s);
     }
-
     return result;
   }
 }
