+++ onto-viewer-toolkit/src/main/java/org/edmcouncil/spec/ontoviewer/toolkit/config/ApplicationConfiguration.java
@@ -2,15 +2,13 @@ package org.edmcouncil.spec.ontoviewer.toolkit.config;
 
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.service.ApplicationConfigurationService;
 import org.edmcouncil.spec.ontoviewer.configloader.configuration.service.YamlMemoryBasedConfigurationService;
-import org.edmcouncil.spec.ontoviewer.configloader.utils.files.FileSystemService;
 import org.edmcouncil.spec.ontoviewer.core.ontology.DetailsManager;
 import org.edmcouncil.spec.ontoviewer.core.ontology.OntologyManager;
-import org.edmcouncil.spec.ontoviewer.core.ontology.data.OwlDataHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.classes.ClassHandler;
 import org.edmcouncil.spec.ontoviewer.core.ontology.data.RestrictionGraphDataHandler;
 import org.edmcouncil.spec.ontoviewer.core.ontology.data.extractor.OwlDataExtractor;
-import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.AnnotationsDataHandler;
-import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.DataHandler;
-import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.IndividualDataHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.data.AnnotationsDataHandler;
+import org.edmcouncil.spec.ontoviewer.core.ontology.data.handler.individual.IndividualDataHelper;
 import org.edmcouncil.spec.ontoviewer.core.ontology.data.label.LabelProvider;
 import org.edmcouncil.spec.ontoviewer.core.ontology.factory.CustomDataFactory;
 import org.springframework.context.annotation.Bean;
@@ -24,13 +22,12 @@ import org.springframework.core.io.ClassPathResource;
 @Import({
     AnnotationsDataHandler.class,
     CustomDataFactory.class,
-    DataHandler.class,
     DetailsManager.class,
-    IndividualDataHandler.class,
+    IndividualDataHelper.class,
     LabelProvider.class,
     OntologyManager.class,
     OwlDataExtractor.class,
-    OwlDataHandler.class,
+    ClassHandler.class,
     RestrictionGraphDataHandler.class,
 })
 @ComponentScan(
