+++ onto-viewer-web-app/config/application_config.yaml
@@ -1,0 +1,8 @@
+application_config:
+  display_license: false
+  display_copyright: false
+  license: 
+    - 'http://purl.org/dc/terms/license'
+  copyright:
+    - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/copyright'
+    - 'https://www.omg.org/spec/Commons/AnnotationVocabulary/copyright'
