+++ onto-viewer-web-app/config/groups_config.yaml
@@ -1,0 +1,106 @@
+groups_config:
+  priority_list:
+    - http://www.w3.org/2000/01/rdf-schema#subClassOf
+    - http://www.w3.org/2000/01/rdf-schema#seeAlso
+  groups:
+    - name: Glossary
+      items:
+        - 'http://www.w3.org/2000/01/rdf-schema#label'
+        - 'http://www.w3.org/2004/02/skos/core#prefLabel'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/hasMaturityLevel'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/synonym'
+        - 'https://www.omg.org/spec/Commons/AnnotationVocabulary/synonym'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/abbreviation'
+        - 'http://www.w3.org/2004/02/skos/core#definition'
+        - 'http://spec.edmcouncil.org/owlnames#mdDefinition'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/definitionOrigin'
+        - 'http://purl.org/dc/terms/source'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/directSource'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/adaptedFrom'
+        - 'http://www.w3.org/2004/02/skos/core#example'
+        - 'http://www.w3.org/2004/02/skos/core#scopeNote'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/usageNote'
+        - 'https://www.omg.org/spec/Commons/AnnotationVocabulary/usageNote'
+        - 'http://www.w3.org/2004/02/skos/core#editorialNote'
+        - 'http://purl.org/dc/terms/title'
+        - 'http://purl.org/dc/terms/abstract'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/keyword'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/fileAbstract'
+        - 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/explanatoryNote'
+        - 'https://www.omg.org/spec/Commons/AnnotationVocabulary/explanatoryNote'
+        - 'http://purl.org/dc/terms/description'
+    - name: This ontology contains
+      items:
+        - '@viewer.internal.clazz'
+        - '@viewer.external.clazz'
+        - '@viewer.internal.objectProperty'
+        - '@viewer.external.objectProperty'
+        - '@viewer.internal.dataProperty'
+        - '@viewer.external.dataProperty'
+        - '@viewer.internal.instance'
+        - '@viewer.external.instance'
+        - '@viewer.internal.annotationProperty'
+        - '@viewer.external.annotationProperty'
+        - '@viewer.internal.dataType'
+        - '@viewer.external.dataType'
+    - name: Ontological characteristic
+      items:
+        - '@viewer.function.direct_sub_data_property'
+        - '@viewer.function.direct_sub_object_property'
+        - '@viewer.function.direct_sub_annotation_property'
+        - '@viewer.function.direct_subclasses'
+        - '@viewer.function.instances'
+        - '@viewer.axiom.EquivalentClasses'
+        - '@viewer.axiom.SubClassOf'
+        - '@viewer.function.anonymous_ancestor'
+        - '@viewer.axiom.DisjointClasses'
+        - '@viewer.axiom.ClassAssertion'
+        - '@viewer.axiom.SameIndividual'
+        - '@viewer.axiom.ObjectPropertyAssertion'
+        - '@viewer.axiom.DataPropertyAssertion'
+        - '@viewer.axiom.SubObjectPropertyOf'
+        - '@viewer.axiom.SubDataPropertyOf'
+        - '@viewer.axiom.SubAnnotationPropertyOf'
+        - '@viewer.axiom.EquivalentObjectProperties'
+        - '@viewer.axiom.EquivalentDataProperties'
+        - '@viewer.axiom.ObjectPropertyDomain'
+        - '@viewer.axiom.DataPropertyDomain'
+        - '@viewer.axiom.ObjectPropertyRange'
+        - '@viewer.axiom.FunctionalObjectProperty'
+        - '@viewer.axiom.DataPropertyRange'
+        - '@viewer.axiom.AnnotationPropertyDomain'
+        - '@viewer.axiom.AnnotationPropertyRangeOf'
+        - '@viewer.axiom.InverseObjectProperties'
+        - 'http://www.w3.org/2002/07/owl#deprecated'
+        - 'http://purl.org/dc/terms/hasPart'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/dependsOn'
+    - name: Meta-information
+      items:
+        - 'http://purl.org/dc/terms/creator'
+        - 'http://purl.org/dc/terms/license'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/copyright'
+        - 'https://www.omg.org/spec/Commons/AnnotationVocabulary/copyright'
+        - 'http://purl.org/dc/terms/contributor'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/contributor'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/moduleAbbreviation'
+        - 'http://purl.org/dc/terms/issued'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/contentLanguage'
+        - 'http://www.w3.org/2004/02/skos/core#changeNote'
+        - 'http://www.w3.org/2004/02/skos/core#note'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/fileAbbreviation'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/filename'
+        - 'http://www.w3.org/2000/01/rdf-schema#comment'
+        - 'http://www.w3.org/2000/01/rdf-schema#isDefinedBy'
+        - 'http://www.w3.org/2000/01/rdf-schema#seeAlso'
+        - 'http://www.w3.org/2004/02/skos/core#historyNote'
+        - 'http://www.omg.org/techprocess/ab/SpecificationMetadata/relatedSpecification'
+        - 'http://purl.org/dc/terms/hasVersion'
+        - 'http://purl.org/dc/terms/references'
+        - 'http://purl.org/dc/terms/modified'
+        - 'http://purl.org/dc/terms/rightsHolder'
+        - 'http://www.w3.org/2002/07/owl#versionInfo'
+        - 'http://purl.org/dc/terms/date'
+        - 'https://spec.edmcouncil.org/idmp/ontology/ISO/ISO21090-HarmonizedDatatypes/isEncodedAs'
+    - name: Usage
+      items:
+        - '@viewer.function.usage_classes'
