+++ onto-viewer-web-app/config/label_config.yaml
@@ -1,0 +1,95 @@
+label_config:
+  display_label: true
+  label_priority: USER_DEFINED
+  force_label_lang: false
+  label_lang: en
+  missing_language_action: FIRST
+  user_default_name_list:
+    - id: 'http://www.w3.org/2000/01/rdf-schema#Literal'
+      name: literal
+    - id: 'http://www.w3.org/2001/XMLSchema#string'
+      name: string
+    - id: 'http://www.omg.org/techprocess/ab/SpecificationMetadata/fileAbbreviation'
+      name: file abbreviation
+    - id: 'http://www.omg.org/techprocess/ab/SpecificationMetadata/filename'
+      name: file name
+    - id: 'http://www.w3.org/2000/01/rdf-schema#isDefinedBy'
+      name: is defined by
+    - id: '@viewer.axiom.EquivalentClasses'
+      name: Equivalent classes (necessary and sufficient criteria)
+    - id: '@viewer.axiom.SubClassOf'
+      name: IS-A restrictions
+    - id: '@viewer.axiom.DisjointClasses'
+      name: Disjoint classes
+    - id: '@viewer.axiom.ClassAssertion'
+      name: Instance classification
+    - id: '@viewer.axiom.SameIndividual'
+      name: same as
+    - id: '@viewer.axiom.ObjectPropertyAssertion'
+      name: Object property assertion
+    - id: '@viewer.axiom.DataPropertyAssertion'
+      name: Data property assertion
+    - id: '@viewer.axiom.DataPropertyDomain'
+      name: Data property domain
+    - id: '@viewer.axiom.SubObjectPropertyOf'
+      name: SubObject property
+    - id: '@viewer.axiom.SubDataPropertyOf'
+      name: SubData property
+    - id: '@viewer.axiom.SubAnnotationPropertyOf'
+      name: SubAnnotation property
+    - id: '@viewer.axiom.EquivalentObjectProperties'
+      name: Equivalent object properties
+    - id: '@viewer.axiom.ObjectPropertyDomain'
+      name: Property domain
+    - id: '@viewer.axiom.ObjectPropertyRange'
+      name: Property range
+    - id: '@viewer.axiom.InverseObjectProperties'
+      name: Inverse object properties
+    - id: '@viewer.axiom.TransitiveObjectProperty'
+      name: Transitive object property
+    - id: '@viewer.axiom.FunctionalObjectProperty'
+      name: Functional object property
+    - id: '@viewer.axiom.DataPropertyRange'
+      name: Data property range
+    - id: '@viewer.axiom.AnnotationPropertyDomain'
+      name: Annotation property domain
+    - id: '@viewer.axiom.AnnotationPropertyRangeOf'
+      name: Annotation property range
+    - id: 'ttps://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/synonym'
+      name: synonym(s)
+    - id: 'http://www.w3.org/2000/01/rdf-schema#seeAlso'
+      name: see also
+    - id: 'http://purl.org/dc/terms/hasPart'
+      name: has part
+    - id: 'http://purl.org/dc/terms/contributor'
+      name: contributor
+    - id: 'http://spec.edmcouncil.org/owlnames#mdDefinition'
+      name: ontological definition
+    - id: 'http://www.w3.org/2004/02/skos/core#definition'
+      name: definition
+    - id: '@viewer.function.instances'
+      name: Instances
+    - id: '@viewer.internal.annotationProperty'
+      name: internal annotation property
+    - id: '@viewer.external.annotationProperty'
+      name: external annotation property
+    - id: '@viewer.stats.numberOfDomain'
+      name: number of domains
+    - id: '@viewer.stats.numberOfModule'
+      name: number of modules
+    - id: '@viewer.stats.numberOfClass'
+      name: number of classes
+    - id: '@viewer.stats.numberOfObjectProperty'
+      name: number of object properties
+    - id: '@viewer.stats.numberOfDataProperty'
+      name: number of data properties
+    - id: '@viewer.stats.numberOfAnnotationProperty'
+      name: number of annotation properties
+    - id: '@viewer.stats.numberOfIndividuals'
+      name: number of individuals
+    - id: '@viewer.stats.numberOfAxiom'
+      name: number of axioms
+    - id: '@viewer.stats.numberOfDatatype'
+      name: number of datatypes
+    - id: '@viewer.stats.numberOfOntologies'
+      name: number of ontologies
