+++ onto-viewer-web-app/config/ontology_config.yaml
@@ -1,0 +1,18 @@
+ontologies:
+  source:
+    - url: 'https://spec.edmcouncil.org/fibo/ontology/MetadataFIBO/'
+    - url: 'https://spec.edmcouncil.org/fibo/ontology/AboutFIBODev/'
+    - path: staticOntologies
+  catalog_path:
+    - ontologies/catalog-v001.xml
+  module_ignore_pattern:
+    - '^(About|Metadata|Load).*'
+    - 'http.*(About|Metadata).*'
+  module_class_iri: 'http://www.omg.org/techprocess/ab/SpecificationMetadata/Module'
+  maturity_level_definition:
+    - iri: 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/Release'
+      label: 'Release'
+    - iri: 'https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/Provisional'
+      label: 'Provisional'
+    - iri: 'https://spec.pistoiaalliance.org/idmp/ontology/META/ChangeManagement/Informative'
+      label: 'Informative'
