+++ onto-viewer-web-app/config/search_config.yaml
@@ -1,0 +1,49 @@
+search_config:
+  search_descriptions:
+    - 'http://www.w3.org/2004/02/skos/core#definition'
+    - 'http://purl.org/dc/terms/abstract'
+  fuzzy_distance: 3
+  reindex_on_start: true
+  find_properties:
+    - label: RDFS Label
+      identifier: rdfs_label
+      iri: http://www.w3.org/2000/01/rdf-schema#label
+    - label: SKOS Definition
+      identifier: skos_definition
+      iri: http://www.w3.org/2004/02/skos/core#definition
+    - label: FIBO Explanatory Note
+      identifier: fibo_explanatoryNote
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/explanatoryNote
+    - label: SKOS Note
+      identifier: skos_note
+      iri: http://www.w3.org/2004/02/skos/core#note
+    - label: FIBO Abbreviation
+      identifier: fibo_abbreviation
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/abbreviation
+    - label: FIBO Common Designation
+      identifier: fibo_commonDesignation
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/commonDesignation
+    - label: PURL Description
+      identifier: purl_description
+      iri: http://purl.org/dc/terms/description
+    - label: FIBO Synonym
+      identifier: fibo_synonym
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/synonym
+    - label: FIBO Preferred Designation
+      identifier: fibo_preferredDesignation
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/preferredDesignation
+    - label: SKOS Example
+      identifier: skos_example
+      iri: http://www.w3.org/2004/02/skos/core#example
+    - label: FIBO Usage Note
+      identifier: fibo_usageNote
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/usageNote
+    - label: SKOS Scope Note
+      identifier: skos_scopeNote
+      iri: http://www.w3.org/2004/02/skos/core#scopeNote
+    - label: FIBO Symbol
+      identifier: fibo_symbol
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/symbol
+    - label: SKOS Alt Label
+      identifier: skos_altLabel
+      iri: https://spec.edmcouncil.org/fibo/ontology/FND/Utilities/AnnotationVocabulary/definitionOrigin
